break out a couple of modules from metamodel.ts especially struct-model.ts and the conglamerate around simple-models. (using an agent)
